requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
